Use getDouble with SQL function AVG()
MySQL that restricts float being queried with getInt caused an
SQLException

All usages of getInt with AVG function replaced with
getDouble cast to int.

Affects issues:
- Fixed #1181

+++ Plan/common/src/main/java/com/djrapitops/plan/storage/database/queries/analysis/PlayerCountQueries.java
@@ -231,7 +231,7 @@ public class PlayerCountQueries {
 
                 @Override
                 public Integer processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getInt("average") : 0;
+                    return set.next() ? (int) set.getDouble("average") : 0;
                 }
             });
         };
@@ -389,7 +389,7 @@ public class PlayerCountQueries {
 
                 @Override
                 public Integer processResults(ResultSet set) throws SQLException {
-                    return set.next() ? set.getInt("average") : 0;
+                    return set.next() ? (int) set.getDouble("average") : 0;
                 }
             });
         };
